Add Pressurized Galaxies
This script changes the progression in Galacticraft to instead use PneumaticCraft's pressure chamber to compress the various metals.  Compressed Iron is also made equivalent in the ore dictionary, as the recipe would end up being the same.

Also adds various other items from Galacticraft to the ore dictionary, as they were not registered.